Updated unit test library to catch2. It seems that newer compilers have introduced type safety on the byte order functions such as htole64 which means they can no longer be used for doubles. So for the moment they have been removed and it is assumed that libentity will only be used on little-endian systems, if that changes in the future then the code will have to be revisited - a static assert has been added to stop compilation on big-endian architectures. A long on 64-bit Windows is not 64-bits like it is on linux, so updated the json codec to explicitly use std::stoll. When compiling with mingw-w64 there appears to be an issue when using std::set<long> with errors regarding ambiguity that just don't occur on other platforms. Unit tests have been updated to use std::set<int> to avoid that ambiguity for now, but when building cross-platform apps this may need to be addressed.

+++ src/test/vref.cpp
@@ -112,7 +112,7 @@ TEST_CASE("vref can map a reference to different types", "[vref]")
 
 	SECTION("can map to a set of simple types")
 	{
-		set<long> items = { 2, 4, 8 };
+		set<int> items	= { 2, 4, 8 };
 		auto amap		= vref<decltype(items)>(items);
 
 		amap.encode(c, dst, "a", stack);
